backend/requirements.txt
# VendeAI - Backend Dependencies

# Flask & Extensions
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.25
mysqlclient==2.2.0
python-dotenv==1.0.0

# Mercado Pago Integration
mercadopago==2.2.1
python-dateutil==2.8.2

# API Requests
requests==2.31.0

# JWT Authentication (opcional)
PyJWT==2.8.0

# Utilities
python-decouple==3.8
